02_InvestSight/requirements.txt
# Microsoft Graph SDK
msgraph-sdk==1.2.0
azure-identity==1.15.0

# 金融數據
yfinance==0.2.31
alpha-vantage==2.3.1

# 新聞數據
feedparser==6.0.10

# 數據處理
pandas==2.1.4
numpy==1.26.2

# 情感分析
textblob==0.17.1

# Excel 處理
openpyxl==3.1.2

# 環境變數
python-dotenv==1.0.0

# 日誌
colorlog==6.8.0

# HTTP 請求
requests==2.31.0

# 測試
pytest==7.4.3
